corrected data types

+++ continuousSafetyGym/bullet_safety_gym/envs/builder.py
@@ -118,13 +118,13 @@ class EnvironmentBuilder(gym.Env):
         # Define limits for observation space and action space
         obs_dim = self.get_observation().shape[0]
         act_dim = self.agent.act_dim
-        o_lim = 1000 * np.ones((obs_dim, ), dtype=np.float32)
-        a_lim = np.ones((act_dim,), dtype=np.float32)
-        self.observation_space = gym.spaces.Box(-o_lim, o_lim, dtype=np.float32)
-        self.action_space = gym.spaces.Box(-a_lim, a_lim, dtype=np.float32)
+        o_lim = 1000 * np.ones((obs_dim, ), dtype=np.float64)
+        a_lim = np.ones((act_dim,), dtype=np.float64)
+        self.observation_space = gym.spaces.Box(-o_lim, o_lim, dtype=np.float64)
+        self.action_space = gym.spaces.Box(-a_lim, a_lim, dtype=np.float64)
 
         cost_len = len(self.task.calculate_cost()['cost'])
-        self.constraint_space = gym.spaces.Box(low=-np.inf, high=+np.inf, shape=(cost_len,), dtype=np.float32)
+        self.constraint_space = gym.spaces.Box(low=-np.inf, high=+np.inf, shape=(cost_len,), dtype=np.float64)
 
 
         # stepping information
